+++ src/classes/controllers/class-admin-controller.php
@@ -14,7 +14,7 @@ if ( ! defined( 'ABSPATH' ) ) {
  *
  * @author      Hugh Lashbrooke
  * @category    Class
- * @package     SeriouslySimplePodcasting/Classes
+ * @package     SeriouslySimplePodcasting/Controllers
  * @since       1.0
  */
 class Admin_Controller {
